chore: move category-data related functions to useFilterContext hook
